requirements.txt
fastapi
uvicorn
python-dotenv
openai
PyPDF2
beautifulsoup4
requests
python-multipart
